code/requirements.txt
streamlit
google-generativeai
faker
pandas
python-dotenv
